backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
opencv-contrib-python
torch
torchvision
numpy
fpdf2
pydantic
